Fix type

+++ menu_analise_combinatoria.h
@@ -1,8 +1,9 @@
 #ifndef MENU_ANALISE_COMBINATORIA_H_INCLUDED
 #define MENU_ANALISE_COMBINATORIA_H_INCLUDED
+#include <stdint.h>
 
 void showMenuAnaliseCombinatoria();
 
-unsigned long long fatorial(unsigned long long n);
+uint64_t fatorial(uint64_t n);
 
 #endif // MENU_ANALISE_COMBINATORIA_H_INCLUDED
